backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
PyJWT
passlib
psycopg2
pandas 
matplotlib
seaborn
asyncpg